Skeleton code for nvcc compilation planner
Summary:
Add a `-_NVCC_DRYRUN_` option to wrap_nvcc.py. If provided, wrap_nvcc.py will generate several files related to how the CUDA file will be compiled (to be implemented):
* A plain dump of the `nvcc -dryrun` output.
* A list of environment variables for the compilation sub-commands to run with.
* A dependency graph of the compilation sub-commands.

This diff adds the skeleton API for this `nvcc_compilation_plan`. This currently runs in addition to the regular monolithic CUDA compilation, so that we still produce a valid C++ output object when `_NVCC_DRYRUN_` is passed. Once the implementation is complete, I'll stop running the mono CUDA compilation in this branch.

Reviewed By: get9

Differential Revision: D69757705

fbshipit-source-id: 5f5c2dc2074c1d9dc6debc20545aaa6d7dcef464