Fix haskell_haddock()
Summary:
D68888352 changed how the C compiler flags were passed to GHC, and
omitted them for `haskell_haddock()`. Sometimes the C flags are needed
for Haddock, in particular when `{-# LANGUAGE CPP #-}` is
enabled. This fixes it, and extends the test case.

It's pretty ugly. Suggestions for doing it in a cleaner way
appreciated.

Reviewed By: pepeiborra

Differential Revision: D70167549

fbshipit-source-id: 77209f9f2522b63a59f3c11bbede7caec33df1d3