Fix regression in duplicate project names
Summary:
I'm guessin the switch to cpp20 caused the targets to have many duplicate dependencies on the same target.
Originally in this case we keep both projects. I looked at a few duplicate project files and noticed that they are exactly the same. Of course this is not guaranteed and we may need to roll this back and figure out something else but I think it's worth the small risk.

Reviewed By: autozimu

Differential Revision: D70203722

fbshipit-source-id: 5bbdd435b8ce046dbb0c259e325074a151503c53