implement standard output for junit/robolectric tests
Summary: Implements a `org.junit.runner.notification.RunListener` and registers it to listen for results from JUnit runs. The listener will be created and registered depending on if TPX sets the `TEST_RESULTS_OUTPUT_FILE`. We control this at the macro level with the `"tpx:supports-test-result-output-spec"` label which we're only adding to the playground targets at this time.

Differential Revision: D70268275

fbshipit-source-id: 45e9d52e53aa190cc40fe525fbfdaae761bedf08