Create a new Kotlin toolchain for fbcode that uses the classic protocol
Summary:
In `fbcode`, targets use the `java_fbcode` toolchain, which specifies `javac` as `fbcode//tools/build/buck/wrappers:javac17.sh`.
Providing `javac` forces `java_library` to build in classic mode, avoiding the use of `javacd`. A similar behavior applies to `kotlin_library`; providing `kotlinc` forces it to use classic build mode.

Before this change, the behavior in `fbcode` was:
* `java_library` built in classic mode.
* `kotlin_library` built in kotlincd mode but used javacd with `fbcode//tools/build/buck/wrappers:javac17.sh` to compile Java code within the Kotlin library.

This was confusing, inconsistent, and difficult to debug. With this change, we introduce a new Kotlin toolchain for fbcode, which provides the Kotlin compiler and ensures the Kotlin library compiles in classic mode.

Reviewed By: IanChilds

Differential Revision: D68627414

fbshipit-source-id: 55c5d415dc13cd2b421bda5fbebdfd18078ca198