Optimise erlang dependency_analyzer VM args
Summary: General escript args were recently optimised via the `_run_escript` function in `erlang_build.bzl`, but the `dependency_analyzer` escript is not executed via that function, so it didn't benefit from this improvement. This change migrates `dependency_analyzer` to be invoked in the standard way via `_run_escript`, which also means it now uses the optimised VM args.

Reviewed By: TheGeorge

Differential Revision: D70552847

fbshipit-source-id: d85bf1d4d3ce33b34343746a9bf5180e028c0797