Native Python for 3.13
Summary:
Make native python work for python 3.13. The primary purpose of this is to unblock 3.13 ft native python testing within meta. A (likely different) version of this will be upstreamed: see https://discuss.python.org/t/c-api-for-initializing-statically-linked-extension-modules/43396 and https://github.com/python/cpython/issues/116146.

In 3.13 private methods were removed and import.c was refactored significantly.

Resolve by patching cpython with exported functions. Calling the functions from our own c code couldn't be done since the function we needed are not exported and if we export those functions we might as well just patch python. The basis of the patch is the same as 3.12

 StaticExtensionLoader is a loader that accesses an unsynchronized map. However since this map is only written at compliation time (generated at build time and never updated) this should be thread safe. The loader has been denoted as not using GIL as documented in https://py-free-threading.github.io/porting/

Comments were added and a small refactor was made for clarity.

This diff also re-enables tests that currently fail w/ 3.13 and ft.

> If youre a future contributor (or possibly me trying to fix 3.14) looking at this being like 'im so confused whats going on' i've documented most of my seplunking [here](https://fb.workplace.com/groups/1059551942441557/permalink/1183541110042639/) and [here](https://fb.workplace.com/groups/1059551942441557/permalink/1177009517362465/)

Reviewed By: itamaro, zsol

Differential Revision: D58795157

fbshipit-source-id: 08df4604859593295c646bf4386ef267052449f1