Don't lose configuration of dependencies
Summary:
https://fburl.com/code/ns21qgb2 will end up reconfiguring all the Rust deps in their default target platform, but that default target platform may not be set correctly in some cases (and even if it is, it may not match the top level target's default platform). Preserve the configuration information when collecting deps instead.

There's a small possibility that this might cause issues, in a case like the following: Imagine we have two top-level targets, `:a` and `:b`, they both depend on `//tp:foo`, and they have different default target platforms. In this case, we would previously end up configuring `//tp:foo` just once, in its default configuration. Now, we will not do that, but it does mean that we may end up with two "versions" of `//tp:foo` in the rust-project.json, one for each configuration. I'm not too worried about this though because:
 1. That might be fine
 2. Multiple adjacent top-level targets with different default target platforms would be pretty unusual.

If it turns into a problem, we could dedupe arbitrarily

Reviewed By: capickett

Differential Revision: D70666329

fbshipit-source-id: 2b3e35284ebd49383dc6f94b63822b059de65550