refactor sdk module path prefixing to use a provided map
Summary: Refactor the SDK module path replacement logic to use a toolchain provided map instead of splitting the logic between the SDK generation script and the build rules. This allows for toolchains with unspecified path components to only include the mappings they provide, and we can make the path replacement logic fail when they are missing.

Reviewed By: drodriguez

Differential Revision: D70504393

fbshipit-source-id: fcb64c3ce631a37c9c38e01d6f4599fc1c9d03c0