Add contents type to Archive representation and introduce virtual archive type
Summary:
Prior to this change, the Archive type Buck uses to represent archives had an external_objects field that was only populated if we used thin archives. It is useful to have access to these objects without needing to opt into thin archives for distributed thin-LTO. To accomplish this, we add another field to the archive representation that specifies what kind of archive we have so we don't have to rely upon checking if external_objects is populated.

We also add a third kind of archive, virtual, in addition to the existing thin and regular archive kinds. This is a regular archive if you consume the artifact itself, but compatible linkers will consume the object files directly between --start-lib --end-lib flags.

Reviewed By: rmaz

Differential Revision: D70339389

fbshipit-source-id: 0a055d8f2258e9eb0582b14904c042b2fb2de9ae