Document and typecheck optimization plan documents
Summary:
The starlark logic in Buck's distributed thin-lto linking rules communicates with the linker during thin-link by writing and reading JSON artifacts. In D66734636 I created classes to document the JSON objects that are used to send information from Buck to the linker. In this diff, I do the same for some of the JSON objects that do the reverse (communicate information produced by the linker back to Buck).

This is a non functional change in preparation for D70101572. Having the fields of the objects documented and consistent makes it easier to understand what is happening.

Reviewed By: rmaz

Differential Revision: D70192548

fbshipit-source-id: 81b7cd8a54c9b98ea566ae222d6a97245f972ce7