Use --force-non-staged with InstrumentationTestRunner pm install for APEX
Summary:
Pass --force-non-staged to the pm install command inside the Android InstrumentationTestRunner. This argument is needed to implement installing the com.meta.xr APEX without having to reboot the device.

After D64108474 landed, it was possible to use adb install multiple times with com.meta.xr APEX and it would do a rebootless install. But buck test is using the pm command internally and was not implementing --force-non-staged.

Now when a test like openxr_nojava_sideload uses APEXES_TO_INSTALL, it forces a non-staged install and the test can run repeatedly. This will fix lots of CI issues we've been seeing with this test.

Reviewed By: scramsby

Differential Revision: D70913453

fbshipit-source-id: 258a1a62eb1dee96e6e4c48ad4167c1e58690f1c