Expose swift incremental compilation option that does not capture incremental output
Summary:
When incremental actions are run on RE, and then run locally, the output of those incremental actions are downloaded prior to local execution. Currently, the incremental output of the Swift compiler is not hermetic and unusable for incrementality across machines.

This means the output from RE cannot be used in a local compilation, and thus the downloaded output is not usable. As such, expose a buckconfig `apple.skip_swift_incremental_outputs`, which when true, we will not capture the output of `.o` and `.swiftdep` files as part of the action. This allows us to not download these files when running locally.

To achieve this, when `skip_swift_incremental_outputs` is set:
1. we do not write the output-file-map json file
2. do not include the swiftdep files as part of the output, only capturing the output files
3. in the swift execution wrapper, write the output-file-map json file with the expected swiftdep and output file mapping

NOTE: This is a temporary workaround until we can get incremental compilations for Swift working on RE as expected

Reviewed By: blackm00n

Differential Revision: D70911279

fbshipit-source-id: cd11a0436705374b366983527daeed7baaf7a805