elp.bxl: load prelude targets first so they can be replaced by wanted targets
Summary:
# Problem

We only want process targets derived from `prelude//erlang/...` if they are dependencies of other targets listed in `.elp.toml` as `included_targets` or `deps_targets`.

To achieve this, we create a single list of targets, each tagged as originating from the prelude, apps or deps.

If we are working on prelude files specifically, they will be listed in the `included_targets` or `deps_targets`.

In `elp.bxl`, we were populating the prelude targets into the list of targets last. This meant that even if they were explicitly listed as `included_targets`, they would be processed as prelude ones and ignored if not referenced elsewhere.

# This Diff

This diff changes the order of populating the target list, adding the prelude ones first, so that they can be replaced by ones tagged `app` or `dep` if they are explicitly wanted.

Reviewed By: michalmuskala, TD5

Differential Revision: D70961137

fbshipit-source-id: 1366bbaf04516375eefb4b99ab2d52b7a9252886