Stop server after disconnect and connection timeout
Summary:
The previous diff is insufficient to prevent zombie processes since there are common scenarios where the installer won't be shut down by buck2.

Buck2 builds and launches the installer in parallel with building the app to install. If the app build fails the installer may have launched without connecting to buck2.
We could make buck always block on connecting and sending a shutdown if it starts but we can't guarantee that buck cleans up correctly since buck could be force killed at all.

There is code in the installer to shut down if the parent process dies but it's broken on Windows and it doesn't seem ideal to require users to run `buck2 kill` if they hit build errors from this.

Reviewed By: navidqar

Differential Revision: D70910673

fbshipit-source-id: 2a67815873a3d18ed9af952e8e19fd62aab1dad5