Fix simulator setup for the iOS 18.3.1 runtime
Summary:
Apple recently added a simulator runtime with a patch version. (ie. '18.3.1' vs. '18.3') This broke simulator creation when using `buck test` because the simulator runtime 'name' was still '18.3', even though the runtime version was '18.3.1', and this caused the resource broker to request the wrong simulator version from `idb_targets`.

This diff updates the resource broker to correctly use the runtime version when creating the command for `idb-targets` to create the simulator.

Differential Revision: D70981636

fbshipit-source-id: a37670d447326025ff8f6a5e7f1b1af3ba818337