Don't use projections in JavaCD/KotlinCD
Summary:
Couple of reasons for this:
(1) Projections are a bit more expensive than just regular outputs, so this reduces memory a bit.
(2) Projections hide that these outputs are actually not being produced, and they are not even really outputs, they're just paths where we know something will be produced. Therefore we don't need to pass them as outputs at all (we should probably move the logic out of the starlark code and into `JavaCDCommand/KotlinCDCommand` tbh, but that's for another day

Reviewed By: jselbo

Differential Revision: D71033561

fbshipit-source-id: d414656fbaa7523c11bb71ef23b9c2de183b2471