add supports_explicit_module_debug_serialization
Summary:
From Xcode 16.2 it is possible to support serialized debug options with explicit modules enabled. This will allow us to improve debugger fidelity by using the clang modules for ObjC type information when debugging Swift.

Add a toolchain and provider attribute to propagate if a toolchain version supports this feature.

Reviewed By: manicaesar

Differential Revision: D71044412

fbshipit-source-id: 7e63b4c8c5954e1a916655adf201622422a22522