Support single file compilation in vsgo
Summary:
Presently, if you select a file in Solution Explorer and hit Ctrl+F7, it uses the default `ClCompile` task, which invokes the system-installed version of cl.exe corresponding to whatever version of Visual Studio you have installed.  It does not go through buck2 or invoke `buck2 build` in any way.  To fix this, we need to do things:

1. We need to replace the `ClCompile` task with a custom one that invokes `Exec` to call buck2.
2. We need to associate the object file sub-target name with each source file so this can be passed to the exec task.

The first is done by Adding a piece of metadata to each `ClCompile` item.  So for example we need to generate something like this for each cpp file:

```
        <ClCompile Include="$(RepoRoot)\A\B\C\Foo\Foo.cpp">
            <BuckObjSubTarget>Foo/Foo.cpp.obj</BuckObjSubTarget>
        </ClCompile>
```

Second, we need to inject a replacement for the `ClCompile` task into every vcxproj.  This means putting the following harcoded snippet of MSBuild after all of the `ClCompile` items.

```
    <Target Name="ClCompile" Condition="'@(ClCompile)' != ''" DependsOnTargets="SelectClCompile">
        <Exec WorkingDirectory="$(RepoRoot)" Command="$(vs_buck_build_command)[objects][%(ClCompile.BuckObjSubTarget)] $(ExtraBuckOptions)"/>
    </Target>
```

One might wonder why this doesn't go inside of `vs_buck_build.props` like the other task replacements.  I was not able to make this work because of the `%(ClCompile.BuckObjSubTarget)` reference.  When I put the task inside of the props file, this would not resolve to anything, so the only way I could get it to work was to put it inside of every vcxproj.

After this PR, hitting Ctrl+F7 on a single file will invoke `buck2 build <target_name>[objects][foo.cpp.obj]`

X-link: https://github.com/facebook/buck2/pull/871

Differential Revision: D71003472

Pulled By: autozimu

fbshipit-source-id: fb3d86ad95916df8e70e3f1ba51fec994fd6cfd4