runtime_dependency_handling introduce first level symlink only setting
Summary: D70927140 changed the behavior of runtime_dependency_handling = "symlink" to symlink transitive deps, which broke some builds. This diff sets the default behavior back to the old behavior for backwards compatibility but makes it such that anyone manually opting in to the behavior with `runtime_dependency_handling = "symlink"` gets the expected transitive symlinking.

Reviewed By: mp3guy

Differential Revision: D71155560

fbshipit-source-id: 97ba0d43fe70e783f1ea50b0ba0c1953ec66d23d