Do not check for jvm_abi_gen_working_dir if the plugin is not used
Summary:
### Context
We are exploring the integration of the Kotlin Incremental Compiler into Buck for the incremental compilation of Kotlin source files. The goal is to improve build speed by only recompiling modified source files within a target, rather than rebuilding the entire target as is currently done.

Recently, I started noticing a continuous increase in compiler non-incremental fallbacks due to the absence of jvmAbiGenWorkingDir. This directory is an output of the jvm-abi-gen plugin, which is used solely for class ABI. Therefore, when the source-only ABI was set as the abi_generation_mode, this directory was never created, leading to a forced non-incremental mode.

### This Diff
This diff checks for jvm_abi_gen_working_dir only if the plugin is being used.

----
### Internal
[Doc](https://docs.google.com/document/d/1-zhGPkmt3I2os-kcIaFk9SvT6odt2AJWswSY-J_6AWs/edit?usp=sharing)|[Unidash](https://fburl.com/unidash/m84c5935)|[Code pointer](https://www.internalfb.com/code/fbsource/[adb2a0b14b76]/fbcode/buck2/prelude/kotlin/kotlincd_jar_creator.bzl?lines=98-107)

Differential Revision: D71127908

fbshipit-source-id: ad7b2337bf22d8d9ef97fb085ba8e7d849229b0f