Dump nvcc sub-commands' env into a file
Summary: With the nvcc -dryrun output, parse the envvars of the nvcc sub-commands and dump them into a file. These envvars are necessary to run the sub-commands with, so that they find the tools they need like cicc for example. `foo.dryrun_dump` contains the actual dryrun dump from building a CUDA file in fbcode, and the top section is the envvars. Added a unit test.

Reviewed By: wenxin0319

Differential Revision: D70540475

fbshipit-source-id: 7a8dd3a7822619725c5b5a49c5efc879f90fe3ce