Fix local simulator setup with deleted runtimes
Summary:
If a user deletes an iOS runtime that's associated with a simulator that was previously created for running tests locally, `idb_targets` returns a simulator without os version information. This causes the resource_broker to throw an exception when parsing the simulator list, making it impossible to run tests locally.

This diff fixes simulator parsing when entries do not contain the expected information.

Reviewed By: jawwad

Differential Revision: D71502120

fbshipit-source-id: 0ea556e21da8fedb3ce6ef21d576a5d4af046d25