Rename KotlinCDScribeLogger to KotlinCDLogger
Summary:
### Context
The Scribe logger was integrated into the Android Build Toolchain to collect data from KotlinCD steps. However, with the recent transition of ABT to open-source, we can no longer include Meta-specific code in the OSS release, as it would cause build failures.

### This Stack
This stack aims to use ServiceLoader to inject the correct implementation based on the repository source (OSS or Meta).

**Final structure**

 {F1976250827}

### This diffs
Renames KotlinCDScribeLogger -> KotlinCDLogger to hide the implementation detail in the interface name.

Differential Revision: D71486547

fbshipit-source-id: 915fa628aefb5e63cea61bfb9f3d886f887b7687