Inject KotlinCDLogger implementation into :command_lib using ServiceLoader
Summary:
### Context
The Scribe logger was integrated into the Android Build Toolchain to collect data from KotlinCD steps. However, with the recent transition of ABT to open-source, we can no longer include Meta-specific code in the OSS release, as it would cause build failures.

### This Stack
This stack aims to use ServiceLoader to inject the correct implementation based on the repository source (OSS or Meta).

**Final structure**

 {F1976250827}

### This diff
Implements the logic for using ServiceLoader to inject KotlinCDLogger into the :command_lib library, instead of having it instantiate the implementation directly.

Reviewed By: jselbo

Differential Revision: D71559745

fbshipit-source-id: 5200621ea734e72b587adbd3c00c5c5c6db27ce1