Create new target for scribe model entity
Summary:
### Context
The Scribe logger was integrated into the Android Build Toolchain to collect data from KotlinCD steps. However, with the recent transition of ABT to open-source, we can no longer include Meta-specific code in the OSS release, as it would cause build failures.

### This Stack
This stack aims to use ServiceLoader to inject the correct implementation based on the repository source (OSS or Meta).

**Final structure**

 {F1976250827}

### This diffs
Creates a new target for the logger's entity object since it will have to be shared through both OSS and non-OSS implementations.

Differential Revision: D71486544

fbshipit-source-id: cb746b48e70db1d9a652262d2e9625ac1f4e35cc