Move KotlinCDLogger interface back to prelude
Summary:
### Context
The Scribe logger was integrated into the Android Build Toolchain to collect data from KotlinCD steps. However, with the recent transition of ABT to open-source, we can no longer include Meta-specific code in the OSS release, as it would cause build failures.

### This Stack
This stack aims to use ServiceLoader to inject the correct implementation based on the repository source (OSS or Meta).

**Final structure**

 {F1976288835}

### This diff
I realized that to maintain a pluggable architecture—where components in the prelude define the contract for implementation, and the client provides it—it's needed to keep such contracts within the prelude itself, rather than outside of it. Therefore, this change moves the KotlinCDLogger interface back to the prelude.

Reviewed By: jselbo

Differential Revision: D71605143

fbshipit-source-id: ae11ecaa87be2b10bad59de4660755a8eb1b1ce4