Use ServiceLoader to load QE2 implementation at runtime
Summary:
### Context
The QE2 framework was integrated into the Android Build Toolchain to conduct an experiment for KSIC. However, with the recent transition of ABT to open-source, we can no longer include Meta-specific code in the OSS release, as it would cause build failures.

### This Stack
This stack implements ServiceLoader-driven injection to select the correct implementation based on the repository source (OSS or Meta), aligning with the existing Scribe implementation pattern.

### This diff
Introduces the necessary components and logic to enable dynamic loading of QE2 implementation at runtime using ServiceLoader.

Reviewed By: RuslanLa

Differential Revision: D71625524

fbshipit-source-id: b96cb7c7f9fccdbd3a73d6540bafab248f9a26a1