Create NoopExperiment* components for OSS builds
Summary:
### Context
The QE2 framework was integrated into the Android Build Toolchain to conduct an experiment for KSIC. However, with the recent transition of ABT to open-source, we can no longer include Meta-specific code in the OSS release, as it would cause build failures.

### This Stack
This stack implements ServiceLoader-driven injection to select the correct implementation based on the repository source (OSS or Meta), aligning with the existing Scribe implementation pattern.

### This diff
Creates NoopExperiment* components in the :noop target to be used as an implementation when building for OSS

Reviewed By: RuslanLa

Differential Revision: D71625528

fbshipit-source-id: 19651caf8ddb24b95200975767f407f24893b211