Implement QE exposure logic with NoopExperimentConfigService component
Summary:
### Context
We are exploring the integration of the Kotlin Incremental Compiler into Buck for the incremental compilation of Kotlin source files. The goal is to improve build speed by only recompiling modified source files within a target, rather than rebuilding the entire target as is currently done.

To ensure a smooth rollout, we would like to experiment with an experimentation framework for KSIC. The primary goal of this framework is to detect any unexpected behavior or error increases.

### This diff
This diff introduces the exposure logic necessary to enable experiment integration.

Important Note: The NoopExperimentConfigService implementation is used, meaning that this change does not impact the current build flow.

Differential Revision: D71720481

fbshipit-source-id: a2b9060e5fdbaf1e3220b646d9a2f0f24ab89ca9