Enable experimentation option for KSIC
Summary:
### Context
We are exploring the integration of the Kotlin Incremental Compiler into Buck for the incremental compilation of Kotlin source files. The goal is to improve build speed by only recompiling modified source files within a target, rather than rebuilding the entire target as is currently done.

To ensure a smooth rollout, we would like to experiment with an experimentation framework for KSIC. The primary goal of this framework is to detect any unexpected behavior or error increases.

### This diff
This diff connects the experimentation framework with the compilation flow.

Note: The experiment is still disabled, so no forking is currently happening

Differential Revision: D71751241

fbshipit-source-id: 09b953599a78f3962b5630cc1c2440b34ef94408