Consider qe2 flag at qe exposure logic
Summary:
### Context
We are exploring the integration of the Kotlin Incremental Compiler into Buck for the incremental compilation of Kotlin source files. The goal is to improve build speed by only recompiling modified source files within a target, rather than rebuilding the entire target as is currently done.

To ensure a smooth rollout, we would like to experiment with an experimentation framework for KSIC. The primary goal of this framework is to detect any unexpected behavior or error increases.

As the first step, I will focus on targeting incremental pure Kotlin targets that do not use additional plugins or annotation processors. Since incremental compilation is enabled on mixed modules and modules using additional plugins or annotation processors, we cannot rely solely on incremental flag as an indicator. Therefore, I have introduced an additional property that indicates whether experimentation is enabled for a given target.

### This diff
This diff introduces the necessary logic to consider the value of the qe2 flag during QE exposure handling.
 ---
Internal
[KSIC Experiment Doc](https://docs.google.com/document/d/1BD1uRf7B3iwBm9sJ6EIabqMFh-sHR6Lw9UefNd6Sw_4/edit?usp=sharing)

Reviewed By: RuslanLa

Differential Revision: D71823305

fbshipit-source-id: 657b32030d97614b0ac7e31a4bca8cebc1585b59