Include Java files for Kotlin Source Change Files
Summary:
### Context
We are exploring the integration of the Kotlin Incremental Compiler into Buck for the incremental compilation of Kotlin source files. The goal is to improve build speed by only recompiling modified source files within a target, rather than rebuilding the entire target as is currently done.

Currently, we allow the compiler to calculate changed files within a target, even though this information is available from Buck as an action input. We can reuse the information from Buck and skip one compiler phase entirely.

### This diff
This diff is a workaround for T209040964 where changing a Java constant won't cause Kotlin usages to recompile.

 **Please note that its implementation is not yet enabled.**
----
Internal
[Doc](https://docs.google.com/document/d/1-zhGPkmt3I2os-kcIaFk9SvT6odt2AJWswSY-J_6AWs/edit?usp=sharing)

Reviewed By: nghuiqin

Differential Revision: D72046347

fbshipit-source-id: c798da09651c28d25d3643bdeac6be72d93b95c7