Hoist skip_proguard flag so it can be provided by the caller
Summary: This was the last property I had to hoist before making proguard run in the java_binary rule. While this isn't android-specific, it's not part of the java_binary rule properties, and the value of it does seem to be android-specific. We can always provide it via java_binary later if we need it there as well, since now it's just passed in rather than fetched from a context.

Reviewed By: hick209

Differential Revision: D72092338

fbshipit-source-id: 99ef3e5f5c3b5410f9d5acf9367f371d9f925a0b